feat: cluster-manager move from core to tools
