+++ .github/workflows/ci-it.yml
@@ -152,7 +152,7 @@ jobs:
 
           - ./internal/tools/orchestrator/...
           - ./internal/tools/pipeline/...
-          - ./internal/tools/cluster-agent/... ./internal/tools/cluster-ops/...
+          - ./internal/tools/cluster-manager/... ./internal/tools/cluster-agent/... ./internal/tools/cluster-ops/...
           - ./internal/tools/gittar/... ./internal/tools/kms/... ./internal/tools/volume-provisioner/...
           - ./internal/tools/monitor/...
           - ./internal/core/...
