+++ internal/tools/cluster-manager/cluster/cluster.go
@@ -25,7 +25,7 @@ import (
 	"github.com/erda-project/erda-proto-go/core/clustermanager/cluster/pb"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/internal/core/cluster-manager/cluster/db"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/cluster/db"
 	"github.com/erda-project/erda/pkg/strutil"
 )
