+++ internal/tools/cluster-manager/cluster/cluster.service.go
@@ -23,8 +23,8 @@ import (
 	"github.com/erda-project/erda-proto-go/core/clustermanager/cluster/pb"
 	orgpb "github.com/erda-project/erda-proto-go/core/org/pb"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/internal/core/cluster-manager/cluster/db"
 	"github.com/erda-project/erda/internal/core/org"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/cluster/db"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/discover"
 	"github.com/erda-project/erda/pkg/http/httputil"
