+++ internal/tools/cluster-manager/cluster/provider.go
@@ -21,8 +21,8 @@ import (
 	"github.com/erda-project/erda-infra/pkg/transport"
 	"github.com/erda-project/erda-proto-go/core/clustermanager/cluster/pb"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/internal/core/cluster-manager/cluster/db"
 	"github.com/erda-project/erda/internal/core/org"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/cluster/db"
 )
 
 type provider struct {
