+++ internal/tools/cluster-manager/dialer/provider.go
@@ -26,8 +26,8 @@ import (
 	clusterpb "github.com/erda-project/erda-proto-go/core/clustermanager/cluster/pb"
 	tokenpb "github.com/erda-project/erda-proto-go/core/token/pb"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/internal/core/cluster-manager/dialer/config"
-	"github.com/erda-project/erda/internal/core/cluster-manager/dialer/server"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/dialer/config"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/dialer/server"
 )
 
 // +provider
