+++ internal/tools/cluster-manager/dialer/server/client_data.go
