+++ internal/tools/cluster-manager/dialer/server/client_data_test.go
