+++ internal/tools/cluster-manager/dialer/server/tunnel-server.go
@@ -42,8 +42,8 @@ import (
 	tokenpb "github.com/erda-project/erda-proto-go/core/token/pb"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/internal/core/cluster-manager/dialer/auth"
-	"github.com/erda-project/erda/internal/core/cluster-manager/dialer/config"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/dialer/auth"
+	"github.com/erda-project/erda/internal/tools/cluster-manager/dialer/config"
 	"github.com/erda-project/erda/pkg/http/httputil"
 )
